Fix ChannelCard reference
